refactor(admin): group graph registry module under admin/graphs
The graph listing page and its composable were stored in view/ and composables/ separately. Move Graphs.vue and useGraphs.ts into one folder so graph browsing logic stays together.
