+++ webui/src/admin/graphs/useGraphs.ts
